Code refactoring, fusion of register modal and login modal
